# Large performance increase from spatial indexes