… pom.xml
… src/org/twak/camp/EdgeSpatialIndex.java
… src/org/twak/camp/HeightCollision.java
… src/org/twak/camp/Skeleton.java