You may want this to live on a separate branch.

Huge performance improvements, from:

1) Slightly faster Edge `collide()` and `isParallel()`.
2) Use a KD-Tree within the previously N^2 loop within `HeightCollision.process()` to quickly find collisions occurring at the same location.
3) And the biggest improvement: Users can now opt to use a constructor that has a `edgeNearestNeighbors` argument. When used, rather than colliding every single corner against every live edge, each corner is **collided only against its k-nearest edges** (queries are backed by an R-Tree). For concave shapes especially, you can get away with a very low k and reduce total time by 100s of times.

The example below has 2400 vertices. With changes 1+2 we go from ~8.5s to ~5.0s. With change 3 (with k=8, which is the min number of nearest neighbour edges needed to produce a valid straight skeleton with this shape) we get to ~40ms!

![image](https://github.com/user-attachments/assets/6a364489-fd1e-4dfe-a5d2-82ebcb53594a)