@twak 2025-02-18
nice work - thanks for this! I'm not setup to test and run this code any more, so it will be a while before I'm able to pull these onto a new branch.

As you say, _generally_ this isn't the correct solution. There are a bunch of fast WSS algorithms that track the speed of a collapsing triangulation (Felkel etc..) that remain fast(er) and correct. For square-only (Manhattan) problems, there are also much much faster algorithms!

What is your application here?